+++ src/commonMain/kotlin/kotools/types/number/StrictlyNegativeDouble.kt
@@ -1,0 +1,14 @@
+package kotools.types.number
+
+import kotools.types.experimental.ExperimentalKotoolsTypesApi
+import kotools.types.internal.ExperimentalSince
+import kotools.types.internal.InternalKotoolsTypesApi
+import kotools.types.internal.KotoolsTypesVersion
+
+/**
+ * Represents a floating-point number of type [Double] that is less than zero.
+ */
+@ExperimentalKotoolsTypesApi
+@ExperimentalSince(KotoolsTypesVersion.Unreleased)
+@OptIn(InternalKotoolsTypesApi::class)
+public class StrictlyNegativeDouble private constructor()